+++ math/sieve-of-eratosthenes/lib/sieve-of-eratosthenes.hpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <vector>
 
-std::vector<int> findPrimes(int maxValue);
+std::vector<uint64_t> findPrimes(uint64_t maxValue);
